Functional mediation analysis under concurrent regression model with point-wise bootstrap confidence interval

This function performs functional mediation regression under the concurrent model with given tuning parameter. Point-wise confidence bands are obtained from bootstrap.

FMA.concurrent.boot(Z, M, Y, intercept = TRUE, basis = NULL, Ld2.basis = NULL,
    basis.type = c("fourier"), nbasis = 3, timeinv = c(0, 1), timegrids = NULL,
    lambda.m = 0.01, lambda.y = 0.01, sims = 1000, boot = TRUE,
    boot.ci.type = c("bca", "perc"), conf.level = 0.95, verbose = TRUE)

Arguments

Z
a data matrix. Z is the treatment trajectory in the mediation analysis. The number of rows is the number of subjects, and the number of columns is the number of measured time points.

M
a data matrix. M is the mediator trajectory in the mediation analysis. The number of rows is the number of subjects, and the number of columns is the number of measured time points.

Y
a data matrix. Y is the outcome trajectory in the mediation analysis. The number of rows is the number of subjects, and the number of columns is the number of measured time points.

intercept
a logic variable. Default is TRUE, an intercept term is included in the regression model.

basis
a data matrix. Basis function used in the functional data analysis. The number of columns is the number of basis function considered. If basis = NULL, Fourier basis functions will be generated.

Ld2.basis
a data matrix. The second derivative of the basis function. The number of columns is the number of basis function considered. If Ld2.basis = NULL, the second derivative of Fourier basis functions will be generated.

basis.type
a character of basis function type. Default is Fourier basis (basis.type = "fourier").

nbasis
an integer, the number of basis function included. If basis is provided, this argument will be ignored.

timeinv
a numeric vector of length two, the time interval considered in the analysis. Default is (0,1).

timegrids
a numeric vector of time grids of measurement. If timegrids = NULL, it is assumed the between measurement time interval is constant.

lambda.m
a numeric value of the tuning parameter in the mediator model.

lambda.y
a numeric value of the tuning parameter in the outcome model.

sims
an integer indicating the number of simulations for inference.

boot
a logical value, indicating whether or not bootstrap should be used. Default is TRUE.

boot.ci.type
a character of confidence interval method. boot.ci.type = "bca" bias corrected confidence interval; boot.ci.type = "perc" percentile confidence interval.

conf.level
a number of significance level. Default is 0.95.

verbose
a logical value, indicating whether print out bootstrap replications.

Details

The concurrent mediation model is $$M(t)=Z(t)\alpha(t)+\epsilon_{1}(t),$$ $$Y(t)=Z(t)\gamma(t)+M(t)\beta(t)+\epsilon_{2}(t),$$ where \(\alpha(t)\), \(\beta(t)\), \(\gamma(t)\) are coefficient curves. The model coefficient curves are estimated by minimizing the penalized \(L_{2}\)-loss.

Value

alpha

a list of output for \(\alpha\) estimate coefficients: the result of the coefficient estimates corresponding to the basis function curve: the point-wise estimate of the coefficient curve

gamma

: a list of output for \(\gamma\) estimate coefficients: the result of the coefficient estimates corresponding to the basis function curve: the point-wise estimate of the coefficient curve

beta

a list of output for \(\beta\) estimate coefficients: the result of the coefficient estimates corresponding to the basis function curve: the point-wise estimate of the coefficient curve

IE

a list of output for indirect effect estimate coefficients: the result of the coefficient estimates corresponding to the basis function curve: the point-wise estimate of the coefficient curve

DE

a list of output for direct effect estimate coefficients: the result of the coefficient estimates corresponding to the basis function curve: the point-wise estimate of the coefficient curve

References

Zhao et al. (2017). Functional Mediation Analysis with an Application to Functional Magnetic Resonance Imaging Data. arXiv preprint arXiv:1805.06923.

Examples

##################################################
# Concurrent functional mediation model
data(env.concurrent)
Z<-get("Z",env.concurrent)
M<-get("M",env.concurrent)
Y<-get("Y",env.concurrent)

# \donttest{
# consider Fourier basis
fit.boot<-FMA.concurrent.boot(Z,M,Y,intercept=FALSE,timeinv=c(0,300))

#> [1] "Bootstrap sample 1"
#> [1] "Bootstrap sample 2"
#> [1] "Bootstrap sample 3"
#> [1] "Bootstrap sample 4"
#> [1] "Bootstrap sample 5"
#> [1] "Bootstrap sample 6"
#> [1] "Bootstrap sample 7"
#> [1] "Bootstrap sample 8"
#> [1] "Bootstrap sample 9"
#> [1] "Bootstrap sample 10"
#> [1] "Bootstrap sample 11"
#> [1] "Bootstrap sample 12"
#> [1] "Bootstrap sample 13"
#> [1] "Bootstrap sample 14"
#> [1] "Bootstrap sample 15"
#> [1] "Bootstrap sample 16"
#> [1] "Bootstrap sample 17"
#> [1] "Bootstrap sample 18"
#> [1] "Bootstrap sample 19"
#> [1] "Bootstrap sample 20"
#> [1] "Bootstrap sample 21"
#> [1] "Bootstrap sample 22"
#> [1] "Bootstrap sample 23"
#> [1] "Bootstrap sample 24"
#> [1] "Bootstrap sample 25"
#> [1] "Bootstrap sample 26"
#> [1] "Bootstrap sample 27"
#> [1] "Bootstrap sample 28"
#> [1] "Bootstrap sample 29"
#> [1] "Bootstrap sample 30"
#> [1] "Bootstrap sample 31"
#> [1] "Bootstrap sample 32"
#> [1] "Bootstrap sample 33"
#> [1] "Bootstrap sample 34"
#> [1] "Bootstrap sample 35"
#> [1] "Bootstrap sample 36"
#> [1] "Bootstrap sample 37"
#> [1] "Bootstrap sample 38"
#> [1] "Bootstrap sample 39"
#> [1] "Bootstrap sample 40"
#> [1] "Bootstrap sample 41"
#> [1] "Bootstrap sample 42"
#> [1] "Bootstrap sample 43"
#> [1] "Bootstrap sample 44"
#> [1] "Bootstrap sample 45"
#> [1] "Bootstrap sample 46"
#> [1] "Bootstrap sample 47"
#> [1] "Bootstrap sample 48"
#> [1] "Bootstrap sample 49"
#> [1] "Bootstrap sample 50"
#> [1] "Bootstrap sample 51"
#> [1] "Bootstrap sample 52"
#> [1] "Bootstrap sample 53"
#> [1] "Bootstrap sample 54"
#> [1] "Bootstrap sample 55"
#> [1] "Bootstrap sample 56"
#> [1] "Bootstrap sample 57"
#> [1] "Bootstrap sample 58"
#> [1] "Bootstrap sample 59"
#> [1] "Bootstrap sample 60"
#> [1] "Bootstrap sample 61"
#> [1] "Bootstrap sample 62"
#> [1] "Bootstrap sample 63"
#> [1] "Bootstrap sample 64"
#> [1] "Bootstrap sample 65"
#> [1] "Bootstrap sample 66"
#> [1] "Bootstrap sample 67"
#> [1] "Bootstrap sample 68"
#> [1] "Bootstrap sample 69"
#> [1] "Bootstrap sample 70"
#> [1] "Bootstrap sample 71"
#> [1] "Bootstrap sample 72"
#> [1] "Bootstrap sample 73"
#> [1] "Bootstrap sample 74"
#> [1] "Bootstrap sample 75"
#> [1] "Bootstrap sample 76"
#> [1] "Bootstrap sample 77"
#> [1] "Bootstrap sample 78"
#> [1] "Bootstrap sample 79"
#> [1] "Bootstrap sample 80"
#> [1] "Bootstrap sample 81"
#> [1] "Bootstrap sample 82"
#> [1] "Bootstrap sample 83"
#> [1] "Bootstrap sample 84"
#> [1] "Bootstrap sample 85"
#> [1] "Bootstrap sample 86"
#> [1] "Bootstrap sample 87"
#> [1] "Bootstrap sample 88"
#> [1] "Bootstrap sample 89"
#> [1] "Bootstrap sample 90"
#> [1] "Bootstrap sample 91"
#> [1] "Bootstrap sample 92"
#> [1] "Bootstrap sample 93"
#> [1] "Bootstrap sample 94"
#> [1] "Bootstrap sample 95"
#> [1] "Bootstrap sample 96"
#> [1] "Bootstrap sample 97"
#> [1] "Bootstrap sample 98"
#> [1] "Bootstrap sample 99"
#> [1] "Bootstrap sample 100"
#> [1] "Bootstrap sample 101"
#> [1] "Bootstrap sample 102"
#> [1] "Bootstrap sample 103"
#> [1] "Bootstrap sample 104"
#> [1] "Bootstrap sample 105"
#> [1] "Bootstrap sample 106"
#> [1] "Bootstrap sample 107"
#> [1] "Bootstrap sample 108"
#> [1] "Bootstrap sample 109"
#> [1] "Bootstrap sample 110"
#> [1] "Bootstrap sample 111"
#> [1] "Bootstrap sample 112"
#> [1] "Bootstrap sample 113"
#> [1] "Bootstrap sample 114"
#> [1] "Bootstrap sample 115"
#> [1] "Bootstrap sample 116"
#> [1] "Bootstrap sample 117"
#> [1] "Bootstrap sample 118"
#> [1] "Bootstrap sample 119"
#> [1] "Bootstrap sample 120"
#> [1] "Bootstrap sample 121"
#> [1] "Bootstrap sample 122"
#> [1] "Bootstrap sample 123"
#> [1] "Bootstrap sample 124"
#> [1] "Bootstrap sample 125"
#> [1] "Bootstrap sample 126"
#> [1] "Bootstrap sample 127"
#> [1] "Bootstrap sample 128"
#> [1] "Bootstrap sample 129"
#> [1] "Bootstrap sample 130"
#> [1] "Bootstrap sample 131"
#> [1] "Bootstrap sample 132"
#> [1] "Bootstrap sample 133"
#> [1] "Bootstrap sample 134"
#> [1] "Bootstrap sample 135"
#> [1] "Bootstrap sample 136"
#> [1] "Bootstrap sample 137"
#> [1] "Bootstrap sample 138"
#> [1] "Bootstrap sample 139"
#> [1] "Bootstrap sample 140"
#> [1] "Bootstrap sample 141"
#> [1] "Bootstrap sample 142"
#> [1] "Bootstrap sample 143"
#> [1] "Bootstrap sample 144"
#> [1] "Bootstrap sample 145"
#> [1] "Bootstrap sample 146"
#> [1] "Bootstrap sample 147"
#> [1] "Bootstrap sample 148"
#> [1] "Bootstrap sample 149"
#> [1] "Bootstrap sample 150"
#> [1] "Bootstrap sample 151"
#> [1] "Bootstrap sample 152"
#> [1] "Bootstrap sample 153"
#> [1] "Bootstrap sample 154"
#> [1] "Bootstrap sample 155"
#> [1] "Bootstrap sample 156"
#> [1] "Bootstrap sample 157"
#> [1] "Bootstrap sample 158"
#> [1] "Bootstrap sample 159"
#> [1] "Bootstrap sample 160"
#> [1] "Bootstrap sample 161"
#> [1] "Bootstrap sample 162"
#> [1] "Bootstrap sample 163"
#> [1] "Bootstrap sample 164"
#> [1] "Bootstrap sample 165"
#> [1] "Bootstrap sample 166"
#> [1] "Bootstrap sample 167"
#> [1] "Bootstrap sample 168"
#> [1] "Bootstrap sample 169"
#> [1] "Bootstrap sample 170"
#> [1] "Bootstrap sample 171"
#> [1] "Bootstrap sample 172"
#> [1] "Bootstrap sample 173"
#> [1] "Bootstrap sample 174"
#> [1] "Bootstrap sample 175"
#> [1] "Bootstrap sample 176"
#> [1] "Bootstrap sample 177"
#> [1] "Bootstrap sample 178"
#> [1] "Bootstrap sample 179"
#> [1] "Bootstrap sample 180"
#> [1] "Bootstrap sample 181"
#> [1] "Bootstrap sample 182"
#> [1] "Bootstrap sample 183"
#> [1] "Bootstrap sample 184"
#> [1] "Bootstrap sample 185"
#> [1] "Bootstrap sample 186"
#> [1] "Bootstrap sample 187"
#> [1] "Bootstrap sample 188"
#> [1] "Bootstrap sample 189"
#> [1] "Bootstrap sample 190"
#> [1] "Bootstrap sample 191"
#> [1] "Bootstrap sample 192"
#> [1] "Bootstrap sample 193"
#> [1] "Bootstrap sample 194"
#> [1] "Bootstrap sample 195"
#> [1] "Bootstrap sample 196"
#> [1] "Bootstrap sample 197"
#> [1] "Bootstrap sample 198"
#> [1] "Bootstrap sample 199"
#> [1] "Bootstrap sample 200"
#> [1] "Bootstrap sample 201"
#> [1] "Bootstrap sample 202"
#> [1] "Bootstrap sample 203"
#> [1] "Bootstrap sample 204"
#> [1] "Bootstrap sample 205"
#> [1] "Bootstrap sample 206"
#> [1] "Bootstrap sample 207"
#> [1] "Bootstrap sample 208"
#> [1] "Bootstrap sample 209"
#> [1] "Bootstrap sample 210"
#> [1] "Bootstrap sample 211"
#> [1] "Bootstrap sample 212"
#> [1] "Bootstrap sample 213"
#> [1] "Bootstrap sample 214"
#> [1] "Bootstrap sample 215"
#> [1] "Bootstrap sample 216"
#> [1] "Bootstrap sample 217"
#> [1] "Bootstrap sample 218"
#> [1] "Bootstrap sample 219"
#> [1] "Bootstrap sample 220"
#> [1] "Bootstrap sample 221"
#> [1] "Bootstrap sample 222"
#> [1] "Bootstrap sample 223"
#> [1] "Bootstrap sample 224"
#> [1] "Bootstrap sample 225"
#> [1] "Bootstrap sample 226"
#> [1] "Bootstrap sample 227"
#> [1] "Bootstrap sample 228"
#> [1] "Bootstrap sample 229"
#> [1] "Bootstrap sample 230"
#> [1] "Bootstrap sample 231"
#> [1] "Bootstrap sample 232"
#> [1] "Bootstrap sample 233"
#> [1] "Bootstrap sample 234"
#> [1] "Bootstrap sample 235"
#> [1] "Bootstrap sample 236"
#> [1] "Bootstrap sample 237"
#> [1] "Bootstrap sample 238"
#> [1] "Bootstrap sample 239"
#> [1] "Bootstrap sample 240"
#> [1] "Bootstrap sample 241"
#> [1] "Bootstrap sample 242"
#> [1] "Bootstrap sample 243"
#> [1] "Bootstrap sample 244"
#> [1] "Bootstrap sample 245"
#> [1] "Bootstrap sample 246"
#> [1] "Bootstrap sample 247"
#> [1] "Bootstrap sample 248"
#> [1] "Bootstrap sample 249"
#> [1] "Bootstrap sample 250"
#> [1] "Bootstrap sample 251"
#> [1] "Bootstrap sample 252"
#> [1] "Bootstrap sample 253"
#> [1] "Bootstrap sample 254"
#> [1] "Bootstrap sample 255"
#> [1] "Bootstrap sample 256"
#> [1] "Bootstrap sample 257"
#> [1] "Bootstrap sample 258"
#> [1] "Bootstrap sample 259"
#> [1] "Bootstrap sample 260"
#> [1] "Bootstrap sample 261"
#> [1] "Bootstrap sample 262"
#> [1] "Bootstrap sample 263"
#> [1] "Bootstrap sample 264"
#> [1] "Bootstrap sample 265"
#> [1] "Bootstrap sample 266"
#> [1] "Bootstrap sample 267"
#> [1] "Bootstrap sample 268"
#> [1] "Bootstrap sample 269"
#> [1] "Bootstrap sample 270"
#> [1] "Bootstrap sample 271"
#> [1] "Bootstrap sample 272"
#> [1] "Bootstrap sample 273"
#> [1] "Bootstrap sample 274"
#> [1] "Bootstrap sample 275"
#> [1] "Bootstrap sample 276"
#> [1] "Bootstrap sample 277"
#> [1] "Bootstrap sample 278"
#> [1] "Bootstrap sample 279"
#> [1] "Bootstrap sample 280"
#> [1] "Bootstrap sample 281"
#> [1] "Bootstrap sample 282"
#> [1] "Bootstrap sample 283"
#> [1] "Bootstrap sample 284"
#> [1] "Bootstrap sample 285"
#> [1] "Bootstrap sample 286"
#> [1] "Bootstrap sample 287"
#> [1] "Bootstrap sample 288"
#> [1] "Bootstrap sample 289"
#> [1] "Bootstrap sample 290"
#> [1] "Bootstrap sample 291"
#> [1] "Bootstrap sample 292"
#> [1] "Bootstrap sample 293"
#> [1] "Bootstrap sample 294"
#> [1] "Bootstrap sample 295"
#> [1] "Bootstrap sample 296"
#> [1] "Bootstrap sample 297"
#> [1] "Bootstrap sample 298"
#> [1] "Bootstrap sample 299"
#> [1] "Bootstrap sample 300"
#> [1] "Bootstrap sample 301"
#> [1] "Bootstrap sample 302"
#> [1] "Bootstrap sample 303"
#> [1] "Bootstrap sample 304"
#> [1] "Bootstrap sample 305"
#> [1] "Bootstrap sample 306"
#> [1] "Bootstrap sample 307"
#> [1] "Bootstrap sample 308"
#> [1] "Bootstrap sample 309"
#> [1] "Bootstrap sample 310"
#> [1] "Bootstrap sample 311"
#> [1] "Bootstrap sample 312"
#> [1] "Bootstrap sample 313"
#> [1] "Bootstrap sample 314"
#> [1] "Bootstrap sample 315"
#> [1] "Bootstrap sample 316"
#> [1] "Bootstrap sample 317"
#> [1] "Bootstrap sample 318"
#> [1] "Bootstrap sample 319"
#> [1] "Bootstrap sample 320"
#> [1] "Bootstrap sample 321"
#> [1] "Bootstrap sample 322"
#> [1] "Bootstrap sample 323"
#> [1] "Bootstrap sample 324"
#> [1] "Bootstrap sample 325"
#> [1] "Bootstrap sample 326"
#> [1] "Bootstrap sample 327"
#> [1] "Bootstrap sample 328"
#> [1] "Bootstrap sample 329"
#> [1] "Bootstrap sample 330"
#> [1] "Bootstrap sample 331"
#> [1] "Bootstrap sample 332"
#> [1] "Bootstrap sample 333"
#> [1] "Bootstrap sample 334"
#> [1] "Bootstrap sample 335"
#> [1] "Bootstrap sample 336"
#> [1] "Bootstrap sample 337"
#> [1] "Bootstrap sample 338"
#> [1] "Bootstrap sample 339"
#> [1] "Bootstrap sample 340"
#> [1] "Bootstrap sample 341"
#> [1] "Bootstrap sample 342"
#> [1] "Bootstrap sample 343"
#> [1] "Bootstrap sample 344"
#> [1] "Bootstrap sample 345"
#> [1] "Bootstrap sample 346"
#> [1] "Bootstrap sample 347"
#> [1] "Bootstrap sample 348"
#> [1] "Bootstrap sample 349"
#> [1] "Bootstrap sample 350"
#> [1] "Bootstrap sample 351"
#> [1] "Bootstrap sample 352"
#> [1] "Bootstrap sample 353"
#> [1] "Bootstrap sample 354"
#> [1] "Bootstrap sample 355"
#> [1] "Bootstrap sample 356"
#> [1] "Bootstrap sample 357"
#> [1] "Bootstrap sample 358"
#> [1] "Bootstrap sample 359"
#> [1] "Bootstrap sample 360"
#> [1] "Bootstrap sample 361"
#> [1] "Bootstrap sample 362"
#> [1] "Bootstrap sample 363"
#> [1] "Bootstrap sample 364"
#> [1] "Bootstrap sample 365"
#> [1] "Bootstrap sample 366"
#> [1] "Bootstrap sample 367"
#> [1] "Bootstrap sample 368"
#> [1] "Bootstrap sample 369"
#> [1] "Bootstrap sample 370"
#> [1] "Bootstrap sample 371"
#> [1] "Bootstrap sample 372"
#> [1] "Bootstrap sample 373"
#> [1] "Bootstrap sample 374"
#> [1] "Bootstrap sample 375"
#> [1] "Bootstrap sample 376"
#> [1] "Bootstrap sample 377"
#> [1] "Bootstrap sample 378"
#> [1] "Bootstrap sample 379"
#> [1] "Bootstrap sample 380"
#> [1] "Bootstrap sample 381"
#> [1] "Bootstrap sample 382"
#> [1] "Bootstrap sample 383"
#> [1] "Bootstrap sample 384"
#> [1] "Bootstrap sample 385"
#> [1] "Bootstrap sample 386"
#> [1] "Bootstrap sample 387"
#> [1] "Bootstrap sample 388"
#> [1] "Bootstrap sample 389"
#> [1] "Bootstrap sample 390"
#> [1] "Bootstrap sample 391"
#> [1] "Bootstrap sample 392"
#> [1] "Bootstrap sample 393"
#> [1] "Bootstrap sample 394"
#> [1] "Bootstrap sample 395"
#> [1] "Bootstrap sample 396"
#> [1] "Bootstrap sample 397"
#> [1] "Bootstrap sample 398"
#> [1] "Bootstrap sample 399"
#> [1] "Bootstrap sample 400"
#> [1] "Bootstrap sample 401"
#> [1] "Bootstrap sample 402"
#> [1] "Bootstrap sample 403"
#> [1] "Bootstrap sample 404"
#> [1] "Bootstrap sample 405"
#> [1] "Bootstrap sample 406"
#> [1] "Bootstrap sample 407"
#> [1] "Bootstrap sample 408"
#> [1] "Bootstrap sample 409"
#> [1] "Bootstrap sample 410"
#> [1] "Bootstrap sample 411"
#> [1] "Bootstrap sample 412"
#> [1] "Bootstrap sample 413"
#> [1] "Bootstrap sample 414"
#> [1] "Bootstrap sample 415"
#> [1] "Bootstrap sample 416"
#> [1] "Bootstrap sample 417"
#> [1] "Bootstrap sample 418"
#> [1] "Bootstrap sample 419"
#> [1] "Bootstrap sample 420"
#> [1] "Bootstrap sample 421"
#> [1] "Bootstrap sample 422"
#> [1] "Bootstrap sample 423"
#> [1] "Bootstrap sample 424"
#> [1] "Bootstrap sample 425"
#> [1] "Bootstrap sample 426"
#> [1] "Bootstrap sample 427"
#> [1] "Bootstrap sample 428"
#> [1] "Bootstrap sample 429"
#> [1] "Bootstrap sample 430"
#> [1] "Bootstrap sample 431"
#> [1] "Bootstrap sample 432"
#> [1] "Bootstrap sample 433"
#> [1] "Bootstrap sample 434"
#> [1] "Bootstrap sample 435"
#> [1] "Bootstrap sample 436"
#> [1] "Bootstrap sample 437"
#> [1] "Bootstrap sample 438"
#> [1] "Bootstrap sample 439"
#> [1] "Bootstrap sample 440"
#> [1] "Bootstrap sample 441"
#> [1] "Bootstrap sample 442"
#> [1] "Bootstrap sample 443"
#> [1] "Bootstrap sample 444"
#> [1] "Bootstrap sample 445"
#> [1] "Bootstrap sample 446"
#> [1] "Bootstrap sample 447"
#> [1] "Bootstrap sample 448"
#> [1] "Bootstrap sample 449"
#> [1] "Bootstrap sample 450"
#> [1] "Bootstrap sample 451"
#> [1] "Bootstrap sample 452"
#> [1] "Bootstrap sample 453"
#> [1] "Bootstrap sample 454"
#> [1] "Bootstrap sample 455"
#> [1] "Bootstrap sample 456"
#> [1] "Bootstrap sample 457"
#> [1] "Bootstrap sample 458"
#> [1] "Bootstrap sample 459"
#> [1] "Bootstrap sample 460"
#> [1] "Bootstrap sample 461"
#> [1] "Bootstrap sample 462"
#> [1] "Bootstrap sample 463"
#> [1] "Bootstrap sample 464"
#> [1] "Bootstrap sample 465"
#> [1] "Bootstrap sample 466"
#> [1] "Bootstrap sample 467"
#> [1] "Bootstrap sample 468"
#> [1] "Bootstrap sample 469"
#> [1] "Bootstrap sample 470"
#> [1] "Bootstrap sample 471"
#> [1] "Bootstrap sample 472"
#> [1] "Bootstrap sample 473"
#> [1] "Bootstrap sample 474"
#> [1] "Bootstrap sample 475"
#> [1] "Bootstrap sample 476"
#> [1] "Bootstrap sample 477"
#> [1] "Bootstrap sample 478"
#> [1] "Bootstrap sample 479"
#> [1] "Bootstrap sample 480"
#> [1] "Bootstrap sample 481"
#> [1] "Bootstrap sample 482"
#> [1] "Bootstrap sample 483"
#> [1] "Bootstrap sample 484"
#> [1] "Bootstrap sample 485"
#> [1] "Bootstrap sample 486"
#> [1] "Bootstrap sample 487"
#> [1] "Bootstrap sample 488"
#> [1] "Bootstrap sample 489"
#> [1] "Bootstrap sample 490"
#> [1] "Bootstrap sample 491"
#> [1] "Bootstrap sample 492"
#> [1] "Bootstrap sample 493"
#> [1] "Bootstrap sample 494"
#> [1] "Bootstrap sample 495"
#> [1] "Bootstrap sample 496"
#> [1] "Bootstrap sample 497"
#> [1] "Bootstrap sample 498"
#> [1] "Bootstrap sample 499"
#> [1] "Bootstrap sample 500"
#> [1] "Bootstrap sample 501"
#> [1] "Bootstrap sample 502"
#> [1] "Bootstrap sample 503"
#> [1] "Bootstrap sample 504"
#> [1] "Bootstrap sample 505"
#> [1] "Bootstrap sample 506"
#> [1] "Bootstrap sample 507"
#> [1] "Bootstrap sample 508"
#> [1] "Bootstrap sample 509"
#> [1] "Bootstrap sample 510"
#> [1] "Bootstrap sample 511"
#> [1] "Bootstrap sample 512"
#> [1] "Bootstrap sample 513"
#> [1] "Bootstrap sample 514"
#> [1] "Bootstrap sample 515"
#> [1] "Bootstrap sample 516"
#> [1] "Bootstrap sample 517"
#> [1] "Bootstrap sample 518"
#> [1] "Bootstrap sample 519"
#> [1] "Bootstrap sample 520"
#> [1] "Bootstrap sample 521"
#> [1] "Bootstrap sample 522"
#> [1] "Bootstrap sample 523"
#> [1] "Bootstrap sample 524"
#> [1] "Bootstrap sample 525"
#> [1] "Bootstrap sample 526"
#> [1] "Bootstrap sample 527"
#> [1] "Bootstrap sample 528"
#> [1] "Bootstrap sample 529"
#> [1] "Bootstrap sample 530"
#> [1] "Bootstrap sample 531"
#> [1] "Bootstrap sample 532"
#> [1] "Bootstrap sample 533"
#> [1] "Bootstrap sample 534"
#> [1] "Bootstrap sample 535"
#> [1] "Bootstrap sample 536"
#> [1] "Bootstrap sample 537"
#> [1] "Bootstrap sample 538"
#> [1] "Bootstrap sample 539"
#> [1] "Bootstrap sample 540"
#> [1] "Bootstrap sample 541"
#> [1] "Bootstrap sample 542"
#> [1] "Bootstrap sample 543"
#> [1] "Bootstrap sample 544"
#> [1] "Bootstrap sample 545"
#> [1] "Bootstrap sample 546"
#> [1] "Bootstrap sample 547"
#> [1] "Bootstrap sample 548"
#> [1] "Bootstrap sample 549"
#> [1] "Bootstrap sample 550"
#> [1] "Bootstrap sample 551"
#> [1] "Bootstrap sample 552"
#> [1] "Bootstrap sample 553"
#> [1] "Bootstrap sample 554"
#> [1] "Bootstrap sample 555"
#> [1] "Bootstrap sample 556"
#> [1] "Bootstrap sample 557"
#> [1] "Bootstrap sample 558"
#> [1] "Bootstrap sample 559"
#> [1] "Bootstrap sample 560"
#> [1] "Bootstrap sample 561"
#> [1] "Bootstrap sample 562"
#> [1] "Bootstrap sample 563"
#> [1] "Bootstrap sample 564"
#> [1] "Bootstrap sample 565"
#> [1] "Bootstrap sample 566"
#> [1] "Bootstrap sample 567"
#> [1] "Bootstrap sample 568"
#> [1] "Bootstrap sample 569"
#> [1] "Bootstrap sample 570"
#> [1] "Bootstrap sample 571"
#> [1] "Bootstrap sample 572"
#> [1] "Bootstrap sample 573"
#> [1] "Bootstrap sample 574"
#> [1] "Bootstrap sample 575"
#> [1] "Bootstrap sample 576"
#> [1] "Bootstrap sample 577"
#> [1] "Bootstrap sample 578"
#> [1] "Bootstrap sample 579"
#> [1] "Bootstrap sample 580"
#> [1] "Bootstrap sample 581"
#> [1] "Bootstrap sample 582"
#> [1] "Bootstrap sample 583"
#> [1] "Bootstrap sample 584"
#> [1] "Bootstrap sample 585"
#> [1] "Bootstrap sample 586"
#> [1] "Bootstrap sample 587"
#> [1] "Bootstrap sample 588"
#> [1] "Bootstrap sample 589"
#> [1] "Bootstrap sample 590"
#> [1] "Bootstrap sample 591"
#> [1] "Bootstrap sample 592"
#> [1] "Bootstrap sample 593"
#> [1] "Bootstrap sample 594"
#> [1] "Bootstrap sample 595"
#> [1] "Bootstrap sample 596"
#> [1] "Bootstrap sample 597"
#> [1] "Bootstrap sample 598"
#> [1] "Bootstrap sample 599"
#> [1] "Bootstrap sample 600"
#> [1] "Bootstrap sample 601"
#> [1] "Bootstrap sample 602"
#> [1] "Bootstrap sample 603"
#> [1] "Bootstrap sample 604"
#> [1] "Bootstrap sample 605"
#> [1] "Bootstrap sample 606"
#> [1] "Bootstrap sample 607"
#> [1] "Bootstrap sample 608"
#> [1] "Bootstrap sample 609"
#> [1] "Bootstrap sample 610"
#> [1] "Bootstrap sample 611"
#> [1] "Bootstrap sample 612"
#> [1] "Bootstrap sample 613"
#> [1] "Bootstrap sample 614"
#> [1] "Bootstrap sample 615"
#> [1] "Bootstrap sample 616"
#> [1] "Bootstrap sample 617"
#> [1] "Bootstrap sample 618"
#> [1] "Bootstrap sample 619"
#> [1] "Bootstrap sample 620"
#> [1] "Bootstrap sample 621"
#> [1] "Bootstrap sample 622"
#> [1] "Bootstrap sample 623"
#> [1] "Bootstrap sample 624"
#> [1] "Bootstrap sample 625"
#> [1] "Bootstrap sample 626"
#> [1] "Bootstrap sample 627"
#> [1] "Bootstrap sample 628"
#> [1] "Bootstrap sample 629"
#> [1] "Bootstrap sample 630"
#> [1] "Bootstrap sample 631"
#> [1] "Bootstrap sample 632"
#> [1] "Bootstrap sample 633"
#> [1] "Bootstrap sample 634"
#> [1] "Bootstrap sample 635"
#> [1] "Bootstrap sample 636"
#> [1] "Bootstrap sample 637"
#> [1] "Bootstrap sample 638"
#> [1] "Bootstrap sample 639"
#> [1] "Bootstrap sample 640"
#> [1] "Bootstrap sample 641"
#> [1] "Bootstrap sample 642"
#> [1] "Bootstrap sample 643"
#> [1] "Bootstrap sample 644"
#> [1] "Bootstrap sample 645"
#> [1] "Bootstrap sample 646"
#> [1] "Bootstrap sample 647"
#> [1] "Bootstrap sample 648"
#> [1] "Bootstrap sample 649"
#> [1] "Bootstrap sample 650"
#> [1] "Bootstrap sample 651"
#> [1] "Bootstrap sample 652"
#> [1] "Bootstrap sample 653"
#> [1] "Bootstrap sample 654"
#> [1] "Bootstrap sample 655"
#> [1] "Bootstrap sample 656"
#> [1] "Bootstrap sample 657"
#> [1] "Bootstrap sample 658"
#> [1] "Bootstrap sample 659"
#> [1] "Bootstrap sample 660"
#> [1] "Bootstrap sample 661"
#> [1] "Bootstrap sample 662"
#> [1] "Bootstrap sample 663"
#> [1] "Bootstrap sample 664"
#> [1] "Bootstrap sample 665"
#> [1] "Bootstrap sample 666"
#> [1] "Bootstrap sample 667"
#> [1] "Bootstrap sample 668"
#> [1] "Bootstrap sample 669"
#> [1] "Bootstrap sample 670"
#> [1] "Bootstrap sample 671"
#> [1] "Bootstrap sample 672"
#> [1] "Bootstrap sample 673"
#> [1] "Bootstrap sample 674"
#> [1] "Bootstrap sample 675"
#> [1] "Bootstrap sample 676"
#> [1] "Bootstrap sample 677"
#> [1] "Bootstrap sample 678"
#> [1] "Bootstrap sample 679"
#> [1] "Bootstrap sample 680"
#> [1] "Bootstrap sample 681"
#> [1] "Bootstrap sample 682"
#> [1] "Bootstrap sample 683"
#> [1] "Bootstrap sample 684"
#> [1] "Bootstrap sample 685"
#> [1] "Bootstrap sample 686"
#> [1] "Bootstrap sample 687"
#> [1] "Bootstrap sample 688"
#> [1] "Bootstrap sample 689"
#> [1] "Bootstrap sample 690"
#> [1] "Bootstrap sample 691"
#> [1] "Bootstrap sample 692"
#> [1] "Bootstrap sample 693"
#> [1] "Bootstrap sample 694"
#> [1] "Bootstrap sample 695"
#> [1] "Bootstrap sample 696"
#> [1] "Bootstrap sample 697"
#> [1] "Bootstrap sample 698"
#> [1] "Bootstrap sample 699"
#> [1] "Bootstrap sample 700"
#> [1] "Bootstrap sample 701"
#> [1] "Bootstrap sample 702"
#> [1] "Bootstrap sample 703"
#> [1] "Bootstrap sample 704"
#> [1] "Bootstrap sample 705"
#> [1] "Bootstrap sample 706"
#> [1] "Bootstrap sample 707"
#> [1] "Bootstrap sample 708"
#> [1] "Bootstrap sample 709"
#> [1] "Bootstrap sample 710"
#> [1] "Bootstrap sample 711"
#> [1] "Bootstrap sample 712"
#> [1] "Bootstrap sample 713"
#> [1] "Bootstrap sample 714"
#> [1] "Bootstrap sample 715"
#> [1] "Bootstrap sample 716"
#> [1] "Bootstrap sample 717"
#> [1] "Bootstrap sample 718"
#> [1] "Bootstrap sample 719"
#> [1] "Bootstrap sample 720"
#> [1] "Bootstrap sample 721"
#> [1] "Bootstrap sample 722"
#> [1] "Bootstrap sample 723"
#> [1] "Bootstrap sample 724"
#> [1] "Bootstrap sample 725"
#> [1] "Bootstrap sample 726"
#> [1] "Bootstrap sample 727"
#> [1] "Bootstrap sample 728"
#> [1] "Bootstrap sample 729"
#> [1] "Bootstrap sample 730"
#> [1] "Bootstrap sample 731"
#> [1] "Bootstrap sample 732"
#> [1] "Bootstrap sample 733"
#> [1] "Bootstrap sample 734"
#> [1] "Bootstrap sample 735"
#> [1] "Bootstrap sample 736"
#> [1] "Bootstrap sample 737"
#> [1] "Bootstrap sample 738"
#> [1] "Bootstrap sample 739"
#> [1] "Bootstrap sample 740"
#> [1] "Bootstrap sample 741"
#> [1] "Bootstrap sample 742"
#> [1] "Bootstrap sample 743"
#> [1] "Bootstrap sample 744"
#> [1] "Bootstrap sample 745"
#> [1] "Bootstrap sample 746"
#> [1] "Bootstrap sample 747"
#> [1] "Bootstrap sample 748"
#> [1] "Bootstrap sample 749"
#> [1] "Bootstrap sample 750"
#> [1] "Bootstrap sample 751"
#> [1] "Bootstrap sample 752"
#> [1] "Bootstrap sample 753"
#> [1] "Bootstrap sample 754"
#> [1] "Bootstrap sample 755"
#> [1] "Bootstrap sample 756"
#> [1] "Bootstrap sample 757"
#> [1] "Bootstrap sample 758"
#> [1] "Bootstrap sample 759"
#> [1] "Bootstrap sample 760"
#> [1] "Bootstrap sample 761"
#> [1] "Bootstrap sample 762"
#> [1] "Bootstrap sample 763"
#> [1] "Bootstrap sample 764"
#> [1] "Bootstrap sample 765"
#> [1] "Bootstrap sample 766"
#> [1] "Bootstrap sample 767"
#> [1] "Bootstrap sample 768"
#> [1] "Bootstrap sample 769"
#> [1] "Bootstrap sample 770"
#> [1] "Bootstrap sample 771"
#> [1] "Bootstrap sample 772"
#> [1] "Bootstrap sample 773"
#> [1] "Bootstrap sample 774"
#> [1] "Bootstrap sample 775"
#> [1] "Bootstrap sample 776"
#> [1] "Bootstrap sample 777"
#> [1] "Bootstrap sample 778"
#> [1] "Bootstrap sample 779"
#> [1] "Bootstrap sample 780"
#> [1] "Bootstrap sample 781"
#> [1] "Bootstrap sample 782"
#> [1] "Bootstrap sample 783"
#> [1] "Bootstrap sample 784"
#> [1] "Bootstrap sample 785"
#> [1] "Bootstrap sample 786"
#> [1] "Bootstrap sample 787"
#> [1] "Bootstrap sample 788"
#> [1] "Bootstrap sample 789"
#> [1] "Bootstrap sample 790"
#> [1] "Bootstrap sample 791"
#> [1] "Bootstrap sample 792"
#> [1] "Bootstrap sample 793"
#> [1] "Bootstrap sample 794"
#> [1] "Bootstrap sample 795"
#> [1] "Bootstrap sample 796"
#> [1] "Bootstrap sample 797"
#> [1] "Bootstrap sample 798"
#> [1] "Bootstrap sample 799"
#> [1] "Bootstrap sample 800"
#> [1] "Bootstrap sample 801"
#> [1] "Bootstrap sample 802"
#> [1] "Bootstrap sample 803"
#> [1] "Bootstrap sample 804"
#> [1] "Bootstrap sample 805"
#> [1] "Bootstrap sample 806"
#> [1] "Bootstrap sample 807"
#> [1] "Bootstrap sample 808"
#> [1] "Bootstrap sample 809"
#> [1] "Bootstrap sample 810"
#> [1] "Bootstrap sample 811"
#> [1] "Bootstrap sample 812"
#> [1] "Bootstrap sample 813"
#> [1] "Bootstrap sample 814"
#> [1] "Bootstrap sample 815"
#> [1] "Bootstrap sample 816"
#> [1] "Bootstrap sample 817"
#> [1] "Bootstrap sample 818"
#> [1] "Bootstrap sample 819"
#> [1] "Bootstrap sample 820"
#> [1] "Bootstrap sample 821"
#> [1] "Bootstrap sample 822"
#> [1] "Bootstrap sample 823"
#> [1] "Bootstrap sample 824"
#> [1] "Bootstrap sample 825"
#> [1] "Bootstrap sample 826"
#> [1] "Bootstrap sample 827"
#> [1] "Bootstrap sample 828"
#> [1] "Bootstrap sample 829"
#> [1] "Bootstrap sample 830"
#> [1] "Bootstrap sample 831"
#> [1] "Bootstrap sample 832"
#> [1] "Bootstrap sample 833"
#> [1] "Bootstrap sample 834"
#> [1] "Bootstrap sample 835"
#> [1] "Bootstrap sample 836"
#> [1] "Bootstrap sample 837"
#> [1] "Bootstrap sample 838"
#> [1] "Bootstrap sample 839"
#> [1] "Bootstrap sample 840"
#> [1] "Bootstrap sample 841"
#> [1] "Bootstrap sample 842"
#> [1] "Bootstrap sample 843"
#> [1] "Bootstrap sample 844"
#> [1] "Bootstrap sample 845"
#> [1] "Bootstrap sample 846"
#> [1] "Bootstrap sample 847"
#> [1] "Bootstrap sample 848"
#> [1] "Bootstrap sample 849"
#> [1] "Bootstrap sample 850"
#> [1] "Bootstrap sample 851"
#> [1] "Bootstrap sample 852"
#> [1] "Bootstrap sample 853"
#> [1] "Bootstrap sample 854"
#> [1] "Bootstrap sample 855"
#> [1] "Bootstrap sample 856"
#> [1] "Bootstrap sample 857"
#> [1] "Bootstrap sample 858"
#> [1] "Bootstrap sample 859"
#> [1] "Bootstrap sample 860"
#> [1] "Bootstrap sample 861"
#> [1] "Bootstrap sample 862"
#> [1] "Bootstrap sample 863"
#> [1] "Bootstrap sample 864"
#> [1] "Bootstrap sample 865"
#> [1] "Bootstrap sample 866"
#> [1] "Bootstrap sample 867"
#> [1] "Bootstrap sample 868"
#> [1] "Bootstrap sample 869"
#> [1] "Bootstrap sample 870"
#> [1] "Bootstrap sample 871"
#> [1] "Bootstrap sample 872"
#> [1] "Bootstrap sample 873"
#> [1] "Bootstrap sample 874"
#> [1] "Bootstrap sample 875"
#> [1] "Bootstrap sample 876"
#> [1] "Bootstrap sample 877"
#> [1] "Bootstrap sample 878"
#> [1] "Bootstrap sample 879"
#> [1] "Bootstrap sample 880"
#> [1] "Bootstrap sample 881"
#> [1] "Bootstrap sample 882"
#> [1] "Bootstrap sample 883"
#> [1] "Bootstrap sample 884"
#> [1] "Bootstrap sample 885"
#> [1] "Bootstrap sample 886"
#> [1] "Bootstrap sample 887"
#> [1] "Bootstrap sample 888"
#> [1] "Bootstrap sample 889"
#> [1] "Bootstrap sample 890"
#> [1] "Bootstrap sample 891"
#> [1] "Bootstrap sample 892"
#> [1] "Bootstrap sample 893"
#> [1] "Bootstrap sample 894"
#> [1] "Bootstrap sample 895"
#> [1] "Bootstrap sample 896"
#> [1] "Bootstrap sample 897"
#> [1] "Bootstrap sample 898"
#> [1] "Bootstrap sample 899"
#> [1] "Bootstrap sample 900"
#> [1] "Bootstrap sample 901"
#> [1] "Bootstrap sample 902"
#> [1] "Bootstrap sample 903"
#> [1] "Bootstrap sample 904"
#> [1] "Bootstrap sample 905"
#> [1] "Bootstrap sample 906"
#> [1] "Bootstrap sample 907"
#> [1] "Bootstrap sample 908"
#> [1] "Bootstrap sample 909"
#> [1] "Bootstrap sample 910"
#> [1] "Bootstrap sample 911"
#> [1] "Bootstrap sample 912"
#> [1] "Bootstrap sample 913"
#> [1] "Bootstrap sample 914"
#> [1] "Bootstrap sample 915"
#> [1] "Bootstrap sample 916"
#> [1] "Bootstrap sample 917"
#> [1] "Bootstrap sample 918"
#> [1] "Bootstrap sample 919"
#> [1] "Bootstrap sample 920"
#> [1] "Bootstrap sample 921"
#> [1] "Bootstrap sample 922"
#> [1] "Bootstrap sample 923"
#> [1] "Bootstrap sample 924"
#> [1] "Bootstrap sample 925"
#> [1] "Bootstrap sample 926"
#> [1] "Bootstrap sample 927"
#> [1] "Bootstrap sample 928"
#> [1] "Bootstrap sample 929"
#> [1] "Bootstrap sample 930"
#> [1] "Bootstrap sample 931"
#> [1] "Bootstrap sample 932"
#> [1] "Bootstrap sample 933"
#> [1] "Bootstrap sample 934"
#> [1] "Bootstrap sample 935"
#> [1] "Bootstrap sample 936"
#> [1] "Bootstrap sample 937"
#> [1] "Bootstrap sample 938"
#> [1] "Bootstrap sample 939"
#> [1] "Bootstrap sample 940"
#> [1] "Bootstrap sample 941"
#> [1] "Bootstrap sample 942"
#> [1] "Bootstrap sample 943"
#> [1] "Bootstrap sample 944"
#> [1] "Bootstrap sample 945"
#> [1] "Bootstrap sample 946"
#> [1] "Bootstrap sample 947"
#> [1] "Bootstrap sample 948"
#> [1] "Bootstrap sample 949"
#> [1] "Bootstrap sample 950"
#> [1] "Bootstrap sample 951"
#> [1] "Bootstrap sample 952"
#> [1] "Bootstrap sample 953"
#> [1] "Bootstrap sample 954"
#> [1] "Bootstrap sample 955"
#> [1] "Bootstrap sample 956"
#> [1] "Bootstrap sample 957"
#> [1] "Bootstrap sample 958"
#> [1] "Bootstrap sample 959"
#> [1] "Bootstrap sample 960"
#> [1] "Bootstrap sample 961"
#> [1] "Bootstrap sample 962"
#> [1] "Bootstrap sample 963"
#> [1] "Bootstrap sample 964"
#> [1] "Bootstrap sample 965"
#> [1] "Bootstrap sample 966"
#> [1] "Bootstrap sample 967"
#> [1] "Bootstrap sample 968"
#> [1] "Bootstrap sample 969"
#> [1] "Bootstrap sample 970"
#> [1] "Bootstrap sample 971"
#> [1] "Bootstrap sample 972"
#> [1] "Bootstrap sample 973"
#> [1] "Bootstrap sample 974"
#> [1] "Bootstrap sample 975"
#> [1] "Bootstrap sample 976"
#> [1] "Bootstrap sample 977"
#> [1] "Bootstrap sample 978"
#> [1] "Bootstrap sample 979"
#> [1] "Bootstrap sample 980"
#> [1] "Bootstrap sample 981"
#> [1] "Bootstrap sample 982"
#> [1] "Bootstrap sample 983"
#> [1] "Bootstrap sample 984"
#> [1] "Bootstrap sample 985"
#> [1] "Bootstrap sample 986"
#> [1] "Bootstrap sample 987"
#> [1] "Bootstrap sample 988"
#> [1] "Bootstrap sample 989"
#> [1] "Bootstrap sample 990"
#> [1] "Bootstrap sample 991"
#> [1] "Bootstrap sample 992"
#> [1] "Bootstrap sample 993"
#> [1] "Bootstrap sample 994"
#> [1] "Bootstrap sample 995"
#> [1] "Bootstrap sample 996"
#> [1] "Bootstrap sample 997"
#> [1] "Bootstrap sample 998"
#> [1] "Bootstrap sample 999"
#> [1] "Bootstrap sample 1000"

# }
##################################################